webscraper/requirements.txt
requests==2.31.0
google-cloud-storage==2.14.0
beautifulsoup4==4.12.3
functions-framework==3.*
sqlalchemy>=1.4.*
cloud-sql-python-connector>=1.6.*
pymysql
google-cloud-secret-manager